+++ src/main/java/com/ururulab/ururu/product/dto/request/ProductOptionRequest.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.product.controller.dto.request;
+package com.ururulab.ururu.product.dto.request;
 
 import com.ururulab.ururu.product.domain.entity.Product;
 import com.ururulab.ururu.product.domain.entity.ProductOption;
@@ -8,8 +8,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
-import static com.ururulab.ururu.product.controller.dto.validation.ProductValidationConstants.*;
-import static com.ururulab.ururu.product.controller.dto.validation.ProductValidationMessages.*;
+import static com.ururulab.ururu.product.dto.validation.ProductValidationConstants.*;
+import static com.ururulab.ururu.product.dto.validation.ProductValidationMessages.*;
 
 public record ProductOptionRequest(
         Long id, // Update시에만 사용 (Create시에는 null)
