+++ src/main/java/com/ururulab/ururu/product/service/CategoryManagementService.java
@@ -10,7 +10,7 @@ import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.ururulab.ururu.product.controller.dto.validation.ProductValidationMessages.*;
+import static com.ururulab.ururu.product.dto.validation.ProductValidationMessages.*;
 
 @Service
 @RequiredArgsConstructor
