+++ src/main/java/com/ururulab/ururu/product/service/ProductService.java
@@ -3,13 +3,13 @@ package com.ururulab.ururu.product.service;
 import com.ururulab.ururu.global.domain.entity.TagCategory;
 import com.ururulab.ururu.global.exception.BusinessException;
 import com.ururulab.ururu.global.exception.error.ErrorCode;
-import com.ururulab.ururu.product.controller.dto.request.ProductImageUploadRequest;
-import com.ururulab.ururu.product.controller.dto.request.ProductNoticeRequest;
-import com.ururulab.ururu.product.controller.dto.request.ProductRequest;
-import com.ururulab.ururu.product.controller.dto.response.*;
+import com.ururulab.ururu.product.dto.request.ProductImageUploadRequest;
+import com.ururulab.ururu.product.dto.request.ProductNoticeRequest;
+import com.ururulab.ururu.product.dto.request.ProductRequest;
 import com.ururulab.ururu.product.domain.entity.*;
 import com.ururulab.ururu.product.domain.entity.enumerated.Status;
 import com.ururulab.ururu.product.domain.repository.*;
+import com.ururulab.ururu.product.dto.response.*;
 import com.ururulab.ururu.product.event.ProductImageUploadEvent;
 import com.ururulab.ururu.product.service.validation.ProductValidator;
 import com.ururulab.ururu.seller.domain.entity.Seller;
