+++ src/main/java/com/ururulab/ururu/product/service/ProductTagService.java
@@ -1,7 +1,7 @@
 package com.ururulab.ururu.product.service;
 
 import com.ururulab.ururu.global.domain.entity.TagCategory;
-import com.ururulab.ururu.product.controller.dto.response.ProductTagResponse;
+import com.ururulab.ururu.product.dto.response.ProductTagResponse;
 import com.ururulab.ururu.product.domain.entity.Product;
 import com.ururulab.ururu.product.domain.entity.ProductTag;
 import com.ururulab.ururu.product.domain.repository.ProductTagRepository;
